Rename c_init.h to _subroutines/c_init.h
