--- c_init.h
@@ -1,57 +1,0 @@
- /*************************************************** 
-    Copyright (C) 2016  Steffen Ochs
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
- ****************************************************/
- 
-
-#include <Ticker.h>               // TIMER
-
-Ticker ticker1;
-Ticker ticker2;
-
-
-//++++++++++++++++++++++++++++++++++++++++++++++++++++++
-// Initialize Variables
-
-bool readyForUpdate = true;
-bool readyForTemp = true;
-
-
-//++++++++++++++++++++++++++++++++++++++++++++++++++++++
-// Initialize Serial
-void set_serial() {
-  Serial.begin(115200);
-  Serial.println();
-  Serial.println();
-}
-
-//++++++++++++++++++++++++++++++++++++++++++++++++++++++
-// Timer Flags
-
-void setReadyForUpdate() {
-  readyForUpdate = true;
-}
-void setReadyForTemp() {
-  readyForTemp = true;
-}
-
-//++++++++++++++++++++++++++++++++++++++++++++++++++++++
-// Initialize Timer
-void set_timer() {
- 
-  ticker1.attach(10, setReadyForUpdate);   // alle 10 sec
-  ticker2.attach(1, setReadyForTemp);   // alle 1 sec
-}
-
